requirements.txt
web3==6.11.1
eth-account==0.10.0
requests==2.31.0
python-dotenv==1.0.0
schedule==1.2.0
pandas==2.1.0
Flask>=3.0.0
flask-cors
